backend/requirements.txt
# Web Framework
flask==3.0.2
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# API Client
requests==2.31.0
python-dotenv==1.0.1

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1 